+++ src/commands/document/document-update.ts
@@ -6,7 +6,7 @@ import { Command } from "commander"
 import { gql } from "../../__codegen__/gql.ts"
 import { readIdsFromStdin } from "../../utils/bulk.ts"
 import { getEditor } from "../../utils/editor.ts"
-import { CliError, NotFoundError, ValidationError, handleError } from "../../utils/errors.ts"
+import { CliError, handleError, NotFoundError, ValidationError } from "../../utils/errors.ts"
 import { getGraphQLClient } from "../../utils/graphql.ts"
 import { isStdinTTY } from "../../utils/runtime.ts"