+++ src/commands/issue/issue-attach.ts
@@ -4,11 +4,11 @@ import { gql } from "../../__codegen__/gql.ts"
 import type { AttachmentCreateInput } from "../../__codegen__/graphql.ts"
 import {
   CliError,
-  NotFoundError,
-  ValidationError,
   handleError,
   isClientError,
   isNotFoundError,
+  NotFoundError,
+  ValidationError,
 } from "../../utils/errors.ts"
 import { getGraphQLClient } from "../../utils/graphql.ts"
 import { shouldShowSpinner } from "../../utils/hyperlink.ts"